Renamed util.string to util.string_helpers so it doesn't conflict with the name of a standard library module.
