+++ overdrive.py
@@ -59,7 +59,7 @@ from util.http import (
     HTTP,
     BadResponseException,
 )
-from util.string import base64
+from util.string_helpers import base64
 from util.worker_pools import RLock
 
 from testing import MockRequestsResponse
